# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ChatCommunicator REST API docs",
    "version": "in dev"
  },
  "paths": {
...
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register user",
        "description": "Route: POST api/users\r\nAuthentication: Not required.",
        "requestBody": {
          "description": "The registration payload containing email, username, and password.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration succeeded and returns the created user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "400": {
            "description": "The request payload failed validation."
          },
          "409": {
            "description": "A user with the same email or username already exists."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get users",
        "description": "Route: GET api/users\r\nAuthorization: Required.\r\nQuery parameters:\r\n  - search: string\r\n  - canBeInvitedByUserId: Guid (required)\r\n            \r\nNote: Currently, only retrieval using the canBeInvitedByUserId filter is supported by the application, as it is the only required scenario.\r\nSearch-only queries may be ignored or have limited behavior depending on the current implementation.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "The free-text search term used to match users (e.g. username or email fragments). Required; pass empty string to return no results or default set (implementation-defined).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "canBeInvitedByUserId",
            "in": "query",
            "description": "Required user identifier. When provided, only users that can receive a friendship invitation from the specified user are returned (e.g. excluding already friends, pending invitations, self).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUsersResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule failure (e.g. invalid GUID format)."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to perform this operation."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
    "/api/users/{userId}/username": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Change user username",
        "description": "Route: PATCH api/users/{userId}/username\r\nAuthorization: Required.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The identifier of the user whose username will be changed.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The payload containing the new username.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Username changed successfully and the new username is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload or business rule violation."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to change this user's username."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "The requested username is already taken."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      }
    },
    "/api/users/{userId}/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Change user password",
        "description": "Route: PATCH api/users/{userId}/password\r\nAuthorization: Required.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The identifier of the user whose password will be changed.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The payload containing the current password and the new password.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully."
          },
          "400": {
            "description": "Invalid request payload or password policy violation."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to change this user's password."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      }
    },
    "/api/users/{userId}/avatar": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Upload user avatar",
        "description": "Route: POST api/users/{userId}/avatar\r\nAuthorization: Required.\r\nConsumes: multipart/form-data.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Identifier of the user whose avatar will be uploaded.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar uploaded successfully."
          },
          "400": {
            "description": "Invalid file provided (missing, zero length, unsupported type, exceeds size limit)."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to upload avatar for this user."
          },
          "404": {
            "description": "User not found."
          },
          "415": {
            "description": "Unsupported media type."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user avatar",
        "description": "Route: DELETE api/users/{userId}/avatar\r\nAuthorization: Required.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The identifier of the user whose avatar will be deleted.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar deleted successfully."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to delete avatar for this user."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Change user avatar",
        "description": "Route: PUT api/users/{userId}/avatar\r\nAuthorization: Required.\r\nConsumes: multipart/form-data.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Identifier of the user whose avatar will be replaced.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar changed successfully."
          },
          "400": {
            "description": "Invalid file provided (missing, zero length, unsupported type, exceeds size limit)."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to change avatar for this user."
          },
          "404": {
            "description": "User not found."
          },
          "415": {
            "description": "Unsupported media type."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
    "/api/users/{userId}/friend-invitations": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user's friendship invitations",
        "description": "Route: GET api/users/{userId}/friend-invitations\r\nAuthorization: Required.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The identifier of the user whose friendship invitations will be retrieved.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friendship invitations retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetFriendshipInvitationResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to access this user's invitations."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      }
    },
    "/api/users/{userId}/chats": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user's chats",
        "description": "Route: GET api/users/{userId}/chats\r\nAuthorization: Required.\r\nQuery parameters:\r\n  - onlyFriends: bool",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The identifier of the user whose chats will be retrieved.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "onlyFriends",
            "in": "query",
            "description": "When true, returns chats that are with friends only; when false, returns all chats available to the user.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chats retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetChatsResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Not authorized to access this user's chats."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "required": [
          "newPassword",
          "oldPassword"
        ],
        "type": "object",
        "properties": {
          "oldPassword": {
            "maxLength": 50,
            "minLength": 6,
            "type": "string"
          },
          "newPassword": {
            "maxLength": 50,
            "minLength": 6,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ChangeUsernameRequest": {
        "required": [
          "newUsername"
        ],
        "type": "object",
        "properties": {
          "newUsername": {
            "maxLength": 25,
            "minLength": 3,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "GetFriendshipInvitationResponse": {
        "type": "object",
        "properties": {
          "friendInvitationId": {
            "type": "string",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "senderUsername": {
            "type": "string",
            "nullable": true
          },
          "senderAvatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUsersResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "isInvited": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "RegisterRequest": {
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 50,
            "minLength": 6,
            "type": "string"
          },
          "username": {
            "maxLength": 25,
            "minLength": 3,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}